requirements.txt
Django>=5.2.5
pip>=25.2
djangorestframework
drf-yasg
